"Version 1.0" This version of the project is now stable, optimized (in terms of speed and ease of usage) and practically aplicable. Folder /src/ contains multiple Python packages.
***1***
/src/PreferenceArticulation
In this package, I created 4 modules/classes.

Class ArticulationExceptions is simply a class which contains all custom-made exceptions which I inherited from the main Python class "Exception". It primarily helped during debugging, but it will throw errors in case invalid arguments are passed to any function in the entire code.

Class Solution is a core class. It makes the whole project more generic. Every solution contains three attributes; x, y and val, where x and y are of class <numpy.ndarray> and val is simply a float. The way this works is that every Solution is characterized by a vector of alternatives (x) and a vector of objective values (y). Since practically every MOO approach used in the project reduces a MO problem to a single objective problem, by somehow using all the criteria and aggregating them into one function. The previous statement is more or less true, the function and way of aggregating can change dinamically, but anyways the point is a scalar which evaluates a solution exists. That scalar is labelled as "val" in class solution.

File BenchmarkObjectives contains two classes, MOO_Problem and MOO_Constraints. Both classes are never meant to be instantiated and are simply a way to keep the code clean. They both contain only static methods. Class MOO_Problem contains 6 methods, BK1, IM1, SCH1, FON, TNK and OSY. Each method is as previously stated static and accepts a <numpy.ndarray> agrument and returns a <numpy.ndarray> as well. The return value contains the objective functions, in an ascending order [f1, f2, ..., fm].
In terms of usage, say I am working on problem FON and I want to evaluate a solution, its simply done as FON(Sol.x), and the vector of objectives is returned. All of these functions are VECTORIZED as much as possible.
The same applies for MOO_Constraints, all its static methods accept an x-vector and return an integer which indicates how many constraints are violated. Sometimes this can only be a single constraint, but sometimes it can be n-many (i.e. the FON constraint only checks if all parameters are within -4<= xi <= 4). To keep the code cleaner, sometimes the constraints of a problem are kept in multiple static methods.

***2***
Package TabuSearch contains for now two important modules.

Module TS implements an abstract TabuSearch. It keeps information about the search, initializes the random seed values etc... Every method is concrete, except one method which must be overridden, and that is the evaluate_solution(self, sol). The reason for this is that every type of preference articulation must specify how solutions are evaluated. The rest remains the same for any type of pref. articulation.

Module weigthing_methods implements the weighting method as a form of a posteriori articulation (can easily be implemented as a priori). It simply implements the "evaluate_solution" method. It is also vectorized and optimized.

By default, every TS is started by the method search(), which as a result saves a *.pickle file containg all information about the search. This way, I can process all the results later and multiple times, without having to conduct the actual search/test again.

***3***
Folder scripts remains as legacy from the previous usage of this project but is not used anymore.

***4***
Package ResultProcessing does a few things.
The script precomputations - as can be inferred, precomputes the search space and pareto front and saves them into the "/data" folder. This way, whenever plots are to be generated, the script generating the plots uses precomputed data to save time.

Script create_txts uses the pickled files to create a "pretty" txt. This is just the easiest way to save all the information of interest in a txt file, which can then easily be used to create a pdf report (for which an entirely different project/package) is used, the reason being that pyPDF2 requirs python2.x

Script create_plots uses the pickled files to create plots, which show the search history (it literarlly plots the search history/path). To speed up, it uses the precomputed data, to show the actual Pareto front and feasible search space every time.

***5***
Package Tests is very modest. Its mainly an indepenent package in case future expansions are to be done. It only contains 1 file for now, which enables the creation and running of n tests of any type.

***6***
Package UnitTesting was used very much during the development. As typos in equations and constraints are very common, I used unittests to check whether the numpy vectorizations were done correctly. For example, I compare the vectorized computations of constraints and benchmark objectives to manually (slowly) computed ones, but certainly correct ones. This package doesn't actually have anything to do with the main project.

******
What remains to be done is simply create two more derived classes of TS, which implement progressive and a priori articulation. That may require some modification to other packages, such as the "create txt" or "create plots" modules, but all those modifications should be very simple and very clean to do. The entire project is now very mature compared to the previous implementations. The performance has been increase by up to 100x, which is thanks to numpy and a few optimization tricks picked up from relevant literature.
An average iteration of TS with a 4-dimensional search space is about 1ms now.